Ricochet unit turnto bug, increased server buffer
Units would turn in the direction of a hit bullet when it was a
ricochet, making them look in weird directions. Now they only turn to
face a bullet when it came directly from a unit.

Bumped the Server message buffer to 5 megs, still craps out though. I
should experiment with UDP for the non critical unit messages